+++ src/IDBCursor.js
@@ -20,7 +20,7 @@ import CFG from './cfg.js';
  * @param {string} valueColumnName
  * @param {boolean} count
  */
-function IDBCursor (range, direction, store, source, keyColumnName, valueColumnName, count, reqSource) {
+function IDBCursor (range, direction, store, source, keyColumnName, valueColumnName, count) {
     // Calling openCursor on an index or objectstore with null is allowed but we treat it as undefined internally
     IDBTransaction.__assertActive(store.transaction);
     if (range === null) {
@@ -49,18 +49,18 @@ function IDBCursor (range, direction, store, source, keyColumnName, valueColumnN
     this.__store = store;
     this.__range = range;
     this.__req = new IDBRequest();
-    this.__req.__source = reqSource !== undefined ? reqSource : this;
+    this.__req.__source = source;
     this.__req.__transaction = this.__store.transaction;
     this.__keyColumnName = keyColumnName;
     this.__valueColumnName = valueColumnName;
     this.__keyOnly = valueColumnName === 'key';
     this.__valueDecoder = this.__keyOnly ? Key : Sca;
     this.__count = count;
-    this.__offset = -1; // Setting this to -1 as continue will set it to 0 anyway
-    this.__lastKeyContinued = undefined; // Used when continuing with a key
+    this.__prefetchedIndex = -1;
     this.__indexSource = util.instanceOf(source, IDBIndex);
     this.__multiEntryIndex = this.__indexSource ? source.multiEntry : false;
     this.__unique = this.direction.includes('unique');
+    this.__sqlDirection = ['prev', 'prevunique'].includes(this.direction) ? 'DESC' : 'ASC';
 
     if (range !== undefined) {
         // Encode the key range and cache the encoded values, so we don't have to re-encode them over and over
@@ -80,6 +80,7 @@ IDBCursor.prototype.__find = function (...args /* key, tx, success, error, recor
 };
 
 IDBCursor.prototype.__findBasic = function (key, tx, success, error, recordsToLoad) {
+    const continueCall = recordsToLoad !== undefined;
     recordsToLoad = recordsToLoad || 1;
 
     const me = this;
@@ -88,20 +89,22 @@ IDBCursor.prototype.__findBasic = function (key, tx, success, error, recordsToLo
     const sqlValues = [];
     sql.push('WHERE', quotedKeyColumnName, 'NOT NULL');
     setSQLForRange(me.__range, quotedKeyColumnName, sql, sqlValues, true, true);
-    if (key !== undefined) {
-        me.__lastKeyContinued = key;
-        me.__offset = 0;
-    }
-    if (me.__lastKeyContinued !== undefined) {
-        sql.push('AND', quotedKeyColumnName, '>= ?');
-        Key.convertValueToKey(me.__lastKeyContinued);
-        sqlValues.push(Key.encode(me.__lastKeyContinued));
-    }
 
     // Determine the ORDER BY direction based on the cursor.
-    if (!me.__count) {
-        const direction = ['prev', 'prevunique'].includes(me.direction) ? 'DESC' : 'ASC';
+    const direction = me.__sqlDirection;
+    const op = direction === 'ASC' ? '>' : '<';
 
+    if (key !== undefined) {
+        sql.push('AND', quotedKeyColumnName, op + '= ?');
+        Key.convertValueToKey(key);
+        sqlValues.push(Key.encode(key));
+    } else if (continueCall && me.__key !== undefined) {
+        sql.push('AND', quotedKeyColumnName, op + ' ?');
+        Key.convertValueToKey(me.__key);
+        sqlValues.push(Key.encode(me.__key));
+    }
+
+    if (!me.__count) {
         // 1. Sort by key
         sql.push('ORDER BY', quotedKeyColumnName, direction); // Todo: Any reason for this first sorting?
 
@@ -116,19 +119,17 @@ IDBCursor.prototype.__findBasic = function (key, tx, success, error, recordsToLo
             // 4. Sort by object store position (if defined and not unique)
             sql.push(',', util.quote(me.__valueColumnName), direction);
         }
-        sql.push('LIMIT', recordsToLoad, 'OFFSET', me.__offset);
+        sql.push('LIMIT', recordsToLoad);
     }
     sql = sql.join(' ');
     CFG.DEBUG && console.log(sql, sqlValues);
 
-    me.__prefetchedData = null;
-    me.__prefetchedIndex = 0;
     tx.executeSql(sql, sqlValues, function (tx, data) {
         if (me.__count) {
             success(undefined, data.rows.length, undefined);
         } else if (data.rows.length > 1) {
-            me.__prefetchedData = data.rows;
             me.__prefetchedIndex = 0;
+            me.__prefetchedData = data.rows;
             CFG.DEBUG && console.log('Preloaded ' + me.__prefetchedData.length + ' records for cursor');
             me.__decode(data.rows.item(0), success);
         } else if (data.rows.length === 1) {
@@ -162,20 +163,22 @@ IDBCursor.prototype.__findMultiEntry = function (key, tx, success, error) {
             sqlValues.push('%' + util.sqlLIKEEscape(me.__range.__lowerCached.slice(0, -1)) + '%');
         }
     }
+
+    // Determine the ORDER BY direction based on the cursor.
+    const direction = me.__sqlDirection;
+    const op = direction === 'ASC' ? '>' : '<';
+
     if (key !== undefined) {
-        me.__lastKeyContinued = key;
-        me.__offset = 0;
-    }
-    if (me.__lastKeyContinued !== undefined) {
-        sql.push('AND', quotedKeyColumnName, '>= ?');
-        Key.convertValueToKey(me.__lastKeyContinued);
-        sqlValues.push(Key.encode(me.__lastKeyContinued));
+        sql.push('AND', quotedKeyColumnName, op + '= ?');
+        Key.convertValueToKey(key);
+        sqlValues.push(Key.encode(key));
+    } else if (me.__key !== undefined) {
+        sql.push('AND', quotedKeyColumnName, op + ' ?');
+        Key.convertValueToKey(me.__key);
+        sqlValues.push(Key.encode(me.__key));
     }
 
     if (!me.__count) {
-        // Determine the ORDER BY direction based on the cursor.
-        const direction = ['prev', 'prevunique'].includes(me.direction) ? 'DESC' : 'ASC';
-
         // 1. Sort by key
         sql.push('ORDER BY', quotedKeyColumnName, direction); // Todo: Any reason for this first sorting?
 
@@ -194,14 +197,20 @@ IDBCursor.prototype.__findMultiEntry = function (key, tx, success, error) {
     sql = sql.join(' ');
     CFG.DEBUG && console.log(sql, sqlValues);
 
-    me.__prefetchedData = null;
-    me.__prefetchedIndex = 0;
     tx.executeSql(sql, sqlValues, function (tx, data) {
-        me.__multiEntryOffset = data.rows.length;
-
         if (data.rows.length > 0) {
+            if (me.__count) { // Avoid caching and other processing below
+                let ct = 0;
+                for (let i = 0; i < data.rows.length; i++) {
+                    const rowItem = data.rows.item(i);
+                    const rowKey = Key.decode(rowItem[me.__keyColumnName], true);
+                    const matches = Key.findMultiEntryMatches(rowKey, me.__range);
+                    ct += matches.length;
+                }
+                success(undefined, ct, undefined);
+                return;
+            }
             const rows = [];
-
             for (let i = 0; i < data.rows.length; i++) {
                 const rowItem = data.rows.item(i);
                 const rowKey = Key.decode(rowItem[me.__keyColumnName], true);
@@ -218,7 +227,6 @@ IDBCursor.prototype.__findMultiEntry = function (key, tx, success, error) {
                     rows.push(clone);
                 }
             }
-
             const reverse = me.direction.indexOf('prev') === 0;
             rows.sort(function (a, b) {
                 if (a.matchingKey.replace('[', 'z') < b.matchingKey.replace('[', 'z')) {
@@ -236,18 +244,15 @@ IDBCursor.prototype.__findMultiEntry = function (key, tx, success, error) {
                 return 0;
             });
 
-            me.__prefetchedData = {
-                data: rows,
-                length: rows.length,
-                item: function (index) {
-                    return this.data[index];
-                }
-            };
-            me.__prefetchedIndex = 0;
-
-            if (me.__count) {
-                success(undefined, rows.length, undefined);
-            } else if (rows.length > 1) {
+            if (rows.length > 1) {
+                me.__prefetchedIndex = 0;
+                me.__prefetchedData = {
+                    data: rows,
+                    length: rows.length,
+                    item: function (index) {
+                        return this.data[index];
+                    }
+                };
                 CFG.DEBUG && console.log('Preloaded ' + me.__prefetchedData.length + ' records for multiEntry cursor');
                 me.__decode(rows[0], success);
             } else if (rows.length === 1) {
@@ -311,12 +316,39 @@ IDBCursor.prototype.__sourceOrEffectiveObjStoreDeleted = function () {
     }
 };
 
-IDBCursor.prototype['continue'] = function (key) {
-    const recordsToPreloadOnContinue = CFG.cursorPreloadPackSize || 100;
+IDBCursor.prototype.__addToCache = function () {
+    this.__prefetchedData = null;
+};
+
+IDBCursor.prototype.__updateCache = function (encKey, encValue) {
     const me = this;
+    if (!me.__prefetchedData) {
+        return;
+    }
+    for (let i = 0; i < me.__prefetchedData.length; i++) {
+        const item = me.__prefetchedData.item(i);
+        if (encKey === item.key) {
+            item.value = encValue;
+            break;
+        }
+    }
+};
+
+IDBCursor.prototype.__deleteFromCache = function (sql, sqlValues) {
+    this.__prefetchedData = null;
+};
+
+IDBCursor.prototype.__clearFromCache = function () {
+    this.__prefetchedData = null;
+};
+
+IDBCursor.prototype.__continue = function (key, advanceContinue) {
+    const me = this;
+    const recordsToPreloadOnContinue = me.__advanceCount || CFG.cursorPreloadPackSize || 100;
+    const advanceState = me.__advanceCount !== undefined;
     IDBTransaction.__assertActive(me.__store.transaction);
     me.__sourceOrEffectiveObjStoreDeleted();
-    if (!me.__gotValue) {
+    if (!me.__gotValue && !advanceContinue) {
         throw createDOMException('InvalidStateError', 'The cursor is being iterated or has iterated past its end.');
     }
     if (key !== undefined) Key.convertValueToKey(key);
@@ -327,26 +359,37 @@ IDBCursor.prototype['continue'] = function (key) {
             (me.direction.includes('next') && cmpResult === -1) ||
             (me.direction.includes('prev') && cmpResult === 1)
         ) {
-            throw createDOMException('DataError', 'Cannot continue the cursor in an unexpected direction');
+            throw createDOMException('DataError', 'Cannot ' + (advanceState ? 'advance' : 'continue') + ' the cursor in an unexpected direction');
         }
     }
 
     me.__gotValue = false;
 
-    me.__store.transaction.__pushToQueue(me.__req, function cursorContinue (tx, args, success, error) {
-        me.__offset++;
-
+    me.__store.transaction.__pushToQueue(me.__req, function cursorContinue (tx, args, success, error, executeNextRequest) {
+        function triggerSuccess (k, val, primKey) {
+            if (advanceState) {
+                if (me.__advanceCount >= 2 && k !== undefined) {
+                    me.__advanceCount--;
+                    me.__key = k;
+                    me.__continue(undefined, true);
+                    executeNextRequest(); // We don't call success yet but do need to advance the transaction queue
+                    return;
+                }
+                me.__advanceCount = undefined;
+            }
+            me.__onsuccess(success)(k, val, primKey);
+        }
         if (me.__prefetchedData) {
             // We have pre-loaded data for the cursor
             me.__prefetchedIndex++;
             if (me.__prefetchedIndex < me.__prefetchedData.length) {
                 me.__decode(me.__prefetchedData.item(me.__prefetchedIndex), function (k, val, primKey) {
-                    function triggerSuccess () {
+                    function checkKey () {
                         if (key !== undefined && k !== key) {
                             cursorContinue(tx, args, success, error);
                             return;
                         }
-                        me.__onsuccess(success)(k, val, primKey);
+                        triggerSuccess(k, val, primKey);
                     }
                     if (me.__unique && !me.__multiEntryIndex) {
                         Sca.encode(val, function (encVal) {
@@ -355,24 +398,31 @@ IDBCursor.prototype['continue'] = function (key) {
                                     cursorContinue(tx, args, success, error);
                                     return;
                                 }
-                                triggerSuccess();
+                                checkKey();
                             });
                         });
                         return;
                     }
-                    triggerSuccess();
+                    checkKey();
                 });
                 return;
             }
         }
 
-        // No pre-fetched data, do query
-        me.__find(key, tx, me.__onsuccess(success), error, recordsToPreloadOnContinue);
+        // No (or not enough) pre-fetched data, do query
+        me.__find(key, tx, triggerSuccess, function (...args) {
+            me.__advanceCount = undefined;
+            error(...args);
+        }, recordsToPreloadOnContinue);
     });
 };
 
+IDBCursor.prototype['continue'] = function (key) {
+    this.__continue(key);
+};
+
 /*
-// Todo:
+// Todo: Implement continuePrimaryKey
 IDBCursor.prototype.continuePrimaryKey = function (key, primaryKey) {};
 */
 
@@ -381,16 +431,10 @@ IDBCursor.prototype.advance = function (count) {
     if (!Number.isFinite(count) || count <= 0) {
         throw new TypeError('Count is invalid - 0 or negative: ' + count);
     }
-    IDBTransaction.__assertActive(me.__store.transaction);
-    me.__sourceOrEffectiveObjStoreDeleted();
-    if (!me.__gotValue) {
-        throw createDOMException('InvalidStateError', 'The cursor is being iterated or has iterated past its end.');
+    if (me.__gotValue) { // Only set the count if not running in error (otherwise will override earlier good advance calls)
+        me.__advanceCount = count;
     }
-    me.__gotValue = false;
-    me.__store.transaction.__pushToQueue(me.__req, function cursorAdvance (tx, args, success, error) {
-        me.__offset += count;
-        me.__find(undefined, tx, me.__onsuccess(success), error);
-    });
+    me.__continue();
 };
 
 IDBCursor.prototype.update = function (valueToUpdate) {
@@ -414,36 +458,40 @@ IDBCursor.prototype.update = function (valueToUpdate) {
     }
     return me.__store.transaction.__addToTransactionQueue(function cursorUpdate (tx, args, success, error) {
         Sca.encode(valueToUpdate, function (encoded) {
-            me.__find(undefined, tx, function (key, value, primaryKey) {
-                const store = me.__store;
-                const params = [encoded];
-                const sql = ['UPDATE', util.escapeStore(store.name), 'SET value = ?'];
-                Key.convertValueToKey(primaryKey);
+            const key = me.key;
+            // const value = me.value; // Todo: Utilize for index?
+            const primaryKey = me.primaryKey;
+            const store = me.__store;
+            // Todo: Fix to run as DELETE and INSERT as per spec (with chance to recreate entry if deleted)
+            const sql = ['UPDATE', util.escapeStore(store.name), 'SET value = ?'];
+            const sqlValues = [encoded];
+            Key.convertValueToKey(primaryKey);
 
-                // Also correct the indexes in the table
-                for (let i = 0; i < store.indexNames.length; i++) {
-                    const index = store.__indexes[store.indexNames[i]];
-                    const indexKey = Key.evaluateKeyPathOnValue(valueToUpdate, index.keyPath, index.multiEntry);
-                    sql.push(',', util.escapeIndex(index.name), '= ?');
-                    params.push(Key.encode(indexKey, index.multiEntry));
+            // Also correct the indexes in the table
+            for (let i = 0; i < store.indexNames.length; i++) {
+                const index = store.__indexes[store.indexNames[i]];
+                const indexKey = Key.evaluateKeyPathOnValue(valueToUpdate, index.keyPath, index.multiEntry);
+                sql.push(',', util.escapeIndex(index.name), '= ?');
+                sqlValues.push(Key.encode(indexKey, index.multiEntry));
+            }
+
+            sql.push('WHERE key = ?');
+            const encodedPrimaryKey = Key.encode(primaryKey);
+            sqlValues.push(encodedPrimaryKey);
+
+            CFG.DEBUG && console.log(sql.join(' '), encoded, key, primaryKey);
+            tx.executeSql(sql.join(' '), sqlValues, function (tx, data) {
+                if (data.rowsAffected === 1) {
+                    me.__store.__cursors.forEach((cursor) => {
+                        cursor.__updateCache(encodedPrimaryKey, encoded);
+                    });
+                    success(key);
+                } else {
+                    error('No rows with key found' + key);
                 }
-
-                sql.push('WHERE key = ?');
-                params.push(Key.encode(primaryKey));
-
-                CFG.DEBUG && console.log(sql.join(' '), encoded, key, primaryKey);
-                tx.executeSql(sql.join(' '), params, function (tx, data) {
-                    me.__prefetchedData = null;
-                    me.__prefetchedIndex = 0;
-                    if (data.rowsAffected === 1) {
-                        success(key);
-                    } else {
-                        error('No rows with key found' + key);
-                    }
-                }, function (tx, data) {
-                    error(data);
-                });
-            }, error);
+            }, function (tx, data) {
+                error(data);
+            });
         });
     }, undefined, me);
 };
@@ -465,11 +513,10 @@ IDBCursor.prototype['delete'] = function () {
             CFG.DEBUG && console.log(sql, key, primaryKey);
             Key.convertValueToKey(primaryKey);
             tx.executeSql(sql, [Key.encode(primaryKey)], function (tx, data) {
-                me.__prefetchedData = null;
-                me.__prefetchedIndex = 0;
                 if (data.rowsAffected === 1) {
-                    // lower the offset or we will miss a row
-                    me.__offset--;
+                    me.__store.__cursors.forEach((cursor) => {
+                        cursor.__deleteFromCache();
+                    });
                     success(undefined);
                 } else {
                     error('No rows with key found' + key);